pyproject.toml
[project]
name = "chatbot1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "langchain>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-groq>=0.3.4",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
